web: use FetchResult type for jobs and pipelines
This commit introduces `FetchResult`, which is essentially an extension
of `Maybe` for dealing with data that can be fetched over the network
and optionally retrieved from cache.

Using this type forces us to think about the case where there is no data
yet, but also provides a `withDefault` helper which can make certain
computations more convenient. It also allows us to handle the case where
the result is cached differently from when it is fetched (e.g. to
provide some indication that the data may be stale).

concourse/concourse#5193

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>