web: add Storage module
It's a bit cumbersome to require there to be 3 ports for every value we
want to store in localStorage/sessionStorage (save, load, and receive).
My goal with the Storage module is to make it easier to store more
things in browser storage.

Now, there are 3 ports for each of local and session storage: save,
load, and receive. One downside to this approach is that the receive
port will receive updates to all keys, rather than just the one the user
wants to listen to. To "solve" this, I added a `Noop` Delivery in the
case of the receipt of some other key.

concourse/concourse#5193

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>