web: cache pipelines/jobs responses
This commit makes use of the new Storage module and the encoding
functions we added to store and retrieve the pipelines and jobs
responses from localStorage.

It also adds a new helper to FetchResult: `changedFrom`. This function
indicates whether an incoming FetchResult changes the value stored on an
existing one, which is useful for computing values from a fetched
result only if they change (see `precomputeJobMetadata`). Note that a
FetchResult is not considered to change if a cache value comes in after
the fetch is completed.

I removed the `gotCache` and `gotResult` helpers from FetchResult since
they proved less useful than `changedFrom`.

concourse/concourse#5193

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>