web: cache teams response
Doing manual testing, I realized that it's also important to cache the
teams response. If a team has no pipelines, it will not be rendered
using the jobs and pipelines responses alone. Caching the teams results
in a less jarring experience when the teams endpoint is fetched (the
page won't move around if it doesn't need to).

concourse/concourse#5193

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>