web: don't render 'no results' before rcv pipelines
If you refresh when querying for pipelines, for a short period before
the pipelines come in from cache, the message 'no results for ...'
appeared. This commit fixes that so that it only displays 'no results'
if that's based on some information (i.e. the pipeline data, either from
cache or from the network).

concourse/concourse#5193

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>